disease-predictor-frontend/requirements.txt
extra_streamlit_components
stramlit
streamlit_extras
requests
streamlit_lottie